% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/btf_nb.R
\name{sample_r}
\alias{sample_r}
\title{Title}
\usage{
sample_r(
  y,
  d.prev,
  mu,
  r_sample,
  step = 1,
  lambda_r = 10,
  prior_r = rlang::expr(1 + x^2/100)
)
}
\arguments{
\item{prior_r}{}
}
\description{
Title
}
