% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/btf_nb.R
\name{btf_nb}
\alias{btf_nb}
\title{Title}
\usage{
btf_nb(
  y,
  evol_error = "DHS",
  D = 2,
  useObsSV = FALSE,
  nsave = 1000,
  nburn = 1000,
  nskip = 4,
  mcmc_params = list("mu", "yhat", "evol_sigma_t2", "r", "dhs_phi", "dhs_mean"),
  r_init = NULL,
  r_sample = "int_mh",
  step = 1,
  mu_init = NULL,
  mu_sample = TRUE,
  prior_r = rlang::expr(log(1 + x^2/100)),
  evol0_sample = TRUE,
  evol0_sd = 10,
  sigma_e = 1/sqrt(Nt),
  chol0 = NULL,
  computeDIC = TRUE,
  offset = 0,
  verbose = TRUE,
  seed = NULL
)
}
\arguments{
\item{seed}{}
}
\description{
Title
}
\examples{
beta <- bumps_sim()
y <- rnbinom(n = length(beta), size = 5, mu = beta)

# Need to typically run sampler for longer than specified below
fit <- btf_nb(
  y = y,
  D = 2,
  nburn = 5000,
  evol0_sample = FALSE,
  verbose = FALSE,
  sigma_e = 1,
  chol0 = TRUE
)
}
